CollectionView Cell implementation for characters
Creating a character cell, RMCharacterCollectionViewCell. The rest of the code in this class is pretty generic, adding views, configuring constraints, prepareForReuse, and also a configure() function with a ViewModel just for our Cell. In that viewModel aka RMCharacterCollectionViewCellViewModel, there is also the function to fetchImage which will later be refactored to a utility class so we can use it on other classes as well. Something to note is that there is a URLError class we can use for when we need built-in foundation error handling.