UI Layout implementatino for Episode Cell and click feature.
Mainly a layout video about the EpisodeCell. We decided to make the episodes parameter from the RMCharacter model public within the RMCharacterDetailViewViewModel, in order for us to access the episode data when we click the cells. Also created a new VC to show the data of a single VC, and we intend to not make another call to fetch that data but just pass it through as we already have that data from earlier when we click a character Cell on the main page.